feat: ARM-242 Removes no longer needed hooks etc inc. refactors
