+++ module/src/components/dateTimeInput/dateTimeInput.component.tsx
@@ -1,5 +1,6 @@
 import { isSameMinute, isValid } from "date-fns";
 import * as React from "react";
+import { isNil } from "lodash";
 
 import { Form } from "../..";
 import { IBindingProps, useBindingState } from "../../hooks/form";
@@ -10,7 +11,6 @@ import {
   NullOrUndefined,
 } from "../../types";
 import { ClassNames, Dates } from "../../utils";
-import { JavaScript } from "../../utils/javascript";
 import { CalendarInput, ICalendarInputProps } from "../calendarInput";
 import { InputWrapper } from "../inputWrapper";
 import { IStatusWrapperProps } from "../statusWrapper";
@@ -143,7 +143,7 @@ export const DateTimeInput = React.forwardRef(
             bind?.addValidationError("Invalid date selection");
             return;
           }
-          const unset = JavaScript.isNullOrUndefined(selectedDateTime);
+          const unset = isNil(selectedDateTime);
           if (
             unset ||
             !isSameMinute(
