--- module/src/components/dropdownButton/index.ts
@@ -1,1 +1,0 @@
-export * from './dropdownButton.component';
