--- module/src/components/emailInput/index.ts
@@ -1,1 +1,0 @@
-export * from './emailInput.component';
