+++ module/src/components/image/image.component.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { HTMLAttributes } from "react";
 
-import { useElementStatus } from "../../hooks/useElementStatus";
+
 import { useHasTimeElapsedSinceMount } from "../../hooks/useHasTimeElapsed";
 import {
   IUseInViewportOptions,
@@ -11,6 +11,7 @@ import { ClassNames } from "../../utils/classNames";
 import { IStatusProps, Status } from "../status";
 
 import "./image.basic.scss";
+import { useElementStatus } from "./image.hooks";
 
 export interface IImageSource {
   /** the url source which will replace the main src if the right conditions are met */
