+++ module/src/components/index.ts
@@ -13,9 +13,7 @@ export * from "./config";
 export * from "./dateTimeInput";
 export * from "./dialog";
 export * from "./dropdown";
-export * from "./dropdownButton";
 export * from "./dropdownItems";
-export * from "./emailInput";
 export * from "./expandable";
 export * from "./icon";
 export * from "./iconButton";
@@ -23,7 +21,6 @@ export * from "./iconWrapper";
 export * from "./image";
 export * from "./input";
 export * from "./inputWrapper";
-export * from "./link";
 export * from "./listBox";
 export * from "./listBoxMulti";
 export * from "./modal";
@@ -37,12 +34,10 @@ export * from "./radioInputList";
 export * from "./rangeInput";
 export * from "./rating";
 export * from "./select";
-export * from "./sideMenu";
 export * from "./spinner";
 export * from "./status";
 export * from "./statusWrapper";
 export * from "./switchInput";
-export * from "./tabControl";
 export * from "./tabSelect";
 export * from "./tag";
 export * from "./tagInput";
