--- module/src/components/link/index.tsx
@@ -1,2 +1,0 @@
-export * from './link.component';
-export * from './link.hooks';
