--- module/src/components/link/link.hooks.ts
@@ -1,4 +1,0 @@
-import { useArmstrongConfig } from '../config';
-
-export const useLocation = () => useArmstrongConfig().routing.location;
-export const useNavigate = () => useArmstrongConfig().routing.navigate;
