+++ module/src/components/radioInput/radioInput.component.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { ArmstrongId, IArmstrongExtendedOption, useGeneratedId } from "../..";
+import { ArmstrongId, IArmstrongExtendedOption } from "../..";
 import { ClassNames } from "../../utils/classNames";
 import { Icon, IconSet, IIcon } from "../icon";
 import { OptionContent } from "../optionContent";
@@ -67,7 +67,7 @@ export const RadioInput = React.forwardRef<HTMLInputElement, IRadioInputProps>(
     },
     ref
   ) => {
-    const generatedId = useGeneratedId("radio_input", inputProps?.id);
+    const generatedId = React.useId();
 
     return (
       <div
