--- module/src/components/sideMenu/index.ts
@@ -1,1 +1,0 @@
-export * from './sideMenu.component';
