--- module/src/components/sideMenu/sideMenu.prototyping.scss
@@ -1,3 +1,0 @@
-.arm-side-menu {
-  border-radius: 0;
-}
