+++ module/src/components/switchInput/switchInput.hooks.ts
@@ -1,6 +1,5 @@
 import * as React from 'react';
-
-import { useEventListener } from './useEventListener';
+import { useEventListener } from '../../hooks/useEventListener';
 
 interface IPosition {
   left: number;
