+++ module/src/components/tabSelect/tabControl/index.ts
