+++ module/src/components/tabSelect/tabControl/tabControl.animated.scss
