+++ module/src/components/tabSelect/tabControl/tabControl.basic.scss
@@ -1,4 +1,4 @@
-@use "../../theme/variables.scss";
+@use "../../../theme/variables.scss";
 
 .arm-tab-control {
   display: flex;
