+++ module/src/components/tabSelect/tabControl/tabControl.component.tsx
@@ -4,12 +4,12 @@ import {
   ArmstrongFCExtensions,
   ArmstrongFCProps,
   ArmstrongFCReturn,
-} from "../../types";
-import { ArmstrongId } from "../../types/core";
-import { IArmstrongExtendedOption } from "../../types/options";
-import { ClassNames } from "../../utils/classNames";
-import { Button } from "../button";
-import { OptionContent } from "../optionContent";
+} from "../../../types";
+import { ArmstrongId } from "../../../types/core";
+import { IArmstrongExtendedOption } from "../../../types/options";
+import { ClassNames } from "../../../utils/classNames";
+import { Button } from "../../button";
+import { OptionContent } from "../../optionContent";
 
 import "./tabControl.basic.scss";
