+++ module/src/hooks/index.ts
@@ -7,14 +7,9 @@ export * from './useArrayState';
 export * from './useBoundingClientRect';
 export * from './useDebounce';
 export * from './useDelayedDependentSwitch';
-export * from './useDidMountEffect';
 export * from './useDidUpdateEffect';
-export * from './useDrag';
-export * from './useElementStatus';
 export * from './useEventListener';
-export * from './useGeneratedId';
 export * from './useHasTimeElapsed';
-export * from './useInfinitePaging';
 export * from './useIntersectionObserver';
 export * from './useInterval';
 export * from './useIsFocused';
@@ -24,10 +19,8 @@ export * from './useMatchMedia';
 export * from './useMutationObserver';
 export * from './useOnScrollToEnd';
 export * from './useOverridableState';
-export * from './usePerformanceObserver';
 export * from './useResizeObserver';
 export * from './useTemporaryState';
 export * from './useThrottle';
 export * from './useTimeout';
-export * from './useWillUnmountEffect';
 export * from './useWindowSize';
