+++ module/src/hooks/useResizeObserver.ts
@@ -2,7 +2,6 @@ import * as React from 'react';
 
 import { Objects } from '../utils';
 import { Globals } from '../utils/globals';
-import { useWillUnMountEffect } from './useWillUnmountEffect';
 
 /**
  * Use an resize observer to fire the passed callback - also cleans up on unmount. Can either be used by just passing in a ref, or by using the functions returned to observe and disconnect
@@ -49,7 +48,7 @@ export function useResizeObserver(
     }
   }, [observe, unobserve]);
 
-  useWillUnMountEffect(disconnect);
+  React.useEffect(() => disconnect, []);
 
   return {
     observer,
