+++ module/src/utils/index.ts
@@ -2,7 +2,6 @@ export * from './arrays';
 export * from './classNames';
 export * from './color';
 export * from './dates';
-export * from './dom';
 export * from './globals';
 export * from './maths';
 export * from './objects';
